storages API
Now PretrainedWeightsProvider takes an Storage instance allowing to share the same API across different storages (e.g. Local of Hugging Face)
