+++ glasses/types.py
@@ -1,0 +1,4 @@
+from typing import Dict
+from torch import Tensor
+
+StateDict = Dict[str, Tensor]
