+++ glasses/utils/weights/storage/__init__.py
@@ -1,0 +1,2 @@
+from .LocalStorage import LocalStorage
+from .HuggingFaceStorage import HuggingFaceStorage
