+++ glasses/utils/weights/storage/hubs/__init__.py
@@ -1,0 +1,1 @@
+from .HFModelHub import HFModelHub
